Simplified dealing with potentials so that problems & opportunities use the same framework (resolves/manifests/nullifies).
Rewrote buy_service as buy_healing: the various offering_service
potentials will each need their own specialized actions to be dealt
with. This is necessary because the delayed effects of buy_service were
visible to the user but not to the system.

Implemented full precondition matching for the surface system, so that
complex preconditions including cross-schema constraints using variables
can work properly (haven't tested this extensively yet).

-Peter Mawhorter